Add QoS 2 message handling test case
