+++ tests/src/receive_spec.cpp
@@ -28,6 +28,7 @@ int test_receive_oversized_message();
 int test_resize_buffer();
 int test_receive_oversized_stream_message();
 int test_receive_qos1();
+int test_receive_qos2();
 
 void reset_callback() {
     callback_called = false;
@@ -330,6 +331,57 @@ int test_receive_qos1() {
     END_IT
 }
 
+int test_receive_qos2() {
+    IT("receives a qos2 message - responds PUBREC then PUBCOMP");
+    reset_callback();
+
+    ShimClient shimClient;
+    shimClient.setAllowConnect(true);
+
+    byte connack[] = {0x20, 0x02, 0x00, 0x00};
+    shimClient.respond(connack, 4);
+
+    PubSubClient client(server, 1883, callback, shimClient);
+    bool rc = client.connect("client_test1");
+    IS_TRUE(rc);
+
+    // QoS 2 PUBLISH from broker (0x34 = MQTTPUBLISH | QoS2 bits)
+    // Fixed header 0x34, remaining length 0x10 (16), topic len 0x0005, topic "topic",
+    // msgId 0x1234, payload "payload"
+    byte publish[] = {0x34, 0x10, 0x0, 0x5, 0x74, 0x6f, 0x70, 0x69, 0x63, 0x12, 0x34, 0x70, 0x61, 0x79, 0x6c, 0x6f, 0x61, 0x64};
+    shimClient.respond(publish, 18);
+
+    // Client must respond with PUBREC (0x50), remaining length 2, msgId 0x1234
+    byte pubrec[] = {0x50, 0x02, 0x12, 0x34};
+    shimClient.expect(pubrec, 4);
+
+    rc = client.loop();
+    IS_TRUE(rc);
+    IS_TRUE(callback_called);
+    IS_TRUE(strcmp(lastTopic, "topic") == 0);
+    IS_TRUE(memcmp(lastPayload, "payload", 7) == 0);
+    IS_TRUE(lastLength == 7);
+    IS_FALSE(shimClient.error());
+
+    reset_callback();
+
+    // Broker sends PUBREL (0x62 = MQTTPUBREL | bit1), remaining length 2, msgId 0x1234
+    byte pubrel[] = {0x62, 0x02, 0x12, 0x34};
+    shimClient.respond(pubrel, 4);
+
+    // Client must respond with PUBCOMP (0x70), remaining length 2, msgId 0x1234
+    byte pubcomp[] = {0x70, 0x02, 0x12, 0x34};
+    shimClient.expect(pubcomp, 4);
+
+    rc = client.loop();
+    IS_TRUE(rc);
+    IS_FALSE(callback_called);  // callback must NOT fire again on PUBREL
+
+    IS_FALSE(shimClient.error());
+
+    END_IT
+}
+
 int main() {
     SUITE("Receive");
     test_receive_callback();
@@ -340,6 +392,7 @@ int main() {
     test_resize_buffer();
     test_receive_oversized_stream_message();
     test_receive_qos1();
+    test_receive_qos2();
 
     FINISH
 }
